requirements.txt
pyyaml
spotipy
pytube
moviepy